fixed possible conversion issue

+++ src/engine/GraphicsEngine.h
@@ -27,7 +27,7 @@ static const SDL_Color SDL_COLOR_GREEN = { 0, 0xFF, 0 };
 static const SDL_Color SDL_COLOR_BLUE = { 0, 0, 0xFF };
 
 inline SDL_Color getRandomColor(int minRGB, int maxRGB) {
-	SDL_Color color = { getRandom(minRGB, maxRGB), getRandom(minRGB, maxRGB), getRandom(minRGB, maxRGB) };
+	SDL_Color color = { (Uint8)getRandom(minRGB, maxRGB), (Uint8)getRandom(minRGB, maxRGB), (Uint8)getRandom(minRGB, maxRGB) };
 	return color;
 }
